bug fixes for streamlines to allow particle sizes to change. also changed some format stuff and fixed some of the casting issues for the new vjobs commit.

+++ VE_SceneGraph/cfdSequence.h
@@ -178,7 +178,7 @@ namespace VE_SceneGraph{
 #endif
 protected:
          cfdSwitch* _lSwitch;
-         int _appFrame;
+         double _appFrame;
          int _lMode;
          int _pMode;
          int _step;
